Move plugin name → instance map to the CLI
Builtin plugins are now imported directly in `dead_cst.cli` and
collected into `_BUILTIN_PLUGINS: dict[str, Plugin]`; the CLI's
`_load_plugin` consults that dict before falling back to the
`dead_cst.plugins` entry-point group for third parties.

Side effects:
- `dead_cst.plugins` no longer re-exports anything from
  `dead_cst.contrib`, and drops `load_plugin` / `BUILTIN_PLUGINS`.
- `dead_cst.contrib` drops its lazy `__getattr__` indirection in
  favour of direct submodule imports.
- `plugins/project_scripts.py` imports `load_toml` from
  `resolvers._core` directly so it doesn't bounce through
  `resolvers/__init__.py` (which transitively reached
  `contrib/__init__.py` and looped back into plugins).
