+++ tests/test_plugins/test_flask.py
@@ -4,7 +4,8 @@ from __future__ import annotations
 
 import pytest
 
-from dead_cst.plugins import DispatchAppPlugin, flask_plugin
+from dead_cst.contrib import flask_plugin
+from dead_cst.plugins import DispatchAppPlugin
 
 
 def test_flask_plugin_marks_route_handlers(build_plugin_graph, reachable_fqnames):
