+++ tests/test_public_api.py
@@ -31,11 +31,7 @@ EXPECTED_PUBLIC_API: dict[str, list[str]] = {
         "write_graph",
     ],
     "dead_cst.plugins": [
-        "BUILTIN_PLUGINS",
-        "CeleryPlugin",
-        "ClickPlugin",
         "DecoratedDeclPlugin",
-        "DiscordPyPlugin",
         "DispatchAppPlugin",
         "DynamicImportFallbackPlugin",
         "EXTERNAL_DIST_PREFIX",
@@ -45,24 +41,14 @@ EXPECTED_PUBLIC_API: dict[str, list[str]] = {
         "InitSubclassPlugin",
         "LiteralListPlugin",
         "MainBlockPlugin",
-        "MockPatchPlugin",
         "ModuleDundersPlugin",
         "Plugin",
         "ProjectScriptsPlugin",
-        "PytestPlugin",
         "STDLIB_PREFIX",
         "SYNTHETIC_PATH_PREFIXES",
-        "ServerConfigPlugin",
         "UNPARSEABLE_PREFIX",
         "UNRESOLVED_PREFIX",
-        "UnittestPlugin",
-        "cyclopts_plugin",
-        "fastapi_plugin",
-        "fastmcp_plugin",
-        "flask_plugin",
-        "load_plugin",
         "simple_name",
-        "typer_plugin",
     ],
     "dead_cst.contrib": [
         "CeleryPlugin",
